requirements.txt
certifi==2020.6.20
cffi==1.14.2
click==7.1.2
cryptography==3.0
decorator==4.4.2
joblib==0.16.0
nltk==3.5
numpy==1.19.1
pycparser==2.20
python-telegram-bot==12.8
regex==2020.7.14
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
tornado==6.0.4
tqdm==4.48.2
